+++ tests/testthat/bad_print/R/print.R
@@ -1,0 +1,3 @@
+print.myclass <- function(x, ...) {
+  cat("myclass object\n")
+}